# Speed up database:getHierarchy: factor >30