Old durations:
665 entries: 0.15729737281799316  and 0.14725184440612793 sec
3340 entries: 2.890997886657715 and 3.463454008102417 sec

New durations:
665 length:0.023437976837158203 and 0.017395496368408203 sec
3340 length: 0.09947752952575684 and 0.08831954002380371 sec